Make aliens more alieny, and draw screen at better time
This commit makes the aliens elliptical instead of rectangular, which
makes them far more alien. It also changes the order in which things
are drawn to the screen, to fix what appeared to be a bug in collision
detection but was actually a case of collisions not being drawn to the
screen. The display now matches up better with the events of the game.